Lots of tests for different border conditions

+++ src/models.rs
@@ -15,11 +15,11 @@ pub mod edge;
 pub mod path;
 pub mod sequence;
 
-use crate::models;
 use crate::models::block::Block;
 use crate::models::edge::Edge;
 use crate::models::path::{all_simple_paths, Path, PathBlock};
 use crate::models::sequence::Sequence;
+use crate::{get_overlap, models};
 
 #[derive(Debug)]
 pub struct Collection {
@@ -359,7 +359,6 @@ impl BlockGroup {
         if ChangeLog::exists(conn, &change.hash) {
             return;
         }
-        println!("change is {path_id} {start} {end} {new_block_id}");
         // todo:
         // 1. get blocks where start-> end overlap
         // 2. split old blocks at boundary points, make new block for left/right side
@@ -378,7 +377,6 @@ impl BlockGroup {
 
         let path = Path::get(conn, path_id);
         let graph = PathBlock::blocks_to_graph(conn, path.id);
-        println!("{path:?} {graph:?}");
         let query = format!("SELECT id, sequence_hash, block_group_id, start, end, strand from block where id in ({block_ids})", block_ids = graph.nodes().map(|k| format!("{k}")).collect::<Vec<_>>().join(","));
         let mut stmt = conn.prepare(&query).unwrap();
         let mut blocks: HashMap<i32, Block> = HashMap::new();
@@ -405,14 +403,22 @@ impl BlockGroup {
         let mut path_end = 0;
         let mut new_edges = vec![];
         let mut previous_block: Option<&Block> = None;
+        let mut loose_connection = false;
         for block_id in &path.blocks {
             let block = blocks.get(block_id).unwrap();
             let block_length = (block.end - block.start);
             path_end += block_length;
 
-            let contains_start = path_start <= start && start < path_end;
-            let contains_end = path_start < end && end < path_end;
-            let overlap = path_start < end && start < path_end;
+            if loose_connection && path_start == end {
+                new_edges.push((Some(new_block_id), Some(block.id)));
+                loose_connection = false;
+            }
+
+            let (contains_start, contains_end, overlap) =
+                get_overlap(path_start, path_end, start, end);
+            println!(
+                "{path_start} {path_end} {start} {end} {contains_start} {contains_end} {overlap}"
+            );
 
             if contains_start && contains_end {
                 // our range is fully contained w/in the block
@@ -452,15 +458,22 @@ impl BlockGroup {
                 // our range is overlapping the end of the block
                 // |----block---|
                 //        |----range---|
-                let (left_block, right_block) = Block::split(
-                    conn,
-                    block,
-                    block.start + start - path_start,
-                    chromosome_index,
-                    phased,
-                )
-                .unwrap();
-                Block::delete(conn, block.id);
+                let split_point = block.start + start - path_start;
+                println!("{block:?} {split_point}");
+                if split_point == block.start {
+                    if let Some(pb) = previous_block {
+                        new_edges.push((Some(pb.id), Some(new_block_id)));
+                    }
+                    // we actually are skipping this block
+                    if path_end >= end {
+                        loose_connection = true;
+                    }
+                } else {
+                    let (left_block, right_block) =
+                        Block::split(conn, block, split_point, chromosome_index, phased).unwrap();
+                    Block::delete(conn, block.id);
+                    new_edges.push((Some(left_block.id), Some(new_block_id)));
+                }
                 // let left_block = Block::create(
                 //     conn,
                 //     &block.sequence_hash,
@@ -474,37 +487,15 @@ impl BlockGroup {
                 // } else {
                 //     new_edges.push((None, Some(left_block.id)));
                 // }
-                new_edges.push((Some(left_block.id), Some(new_block_id)));
             } else if contains_end {
                 // our range is overlapping the beginning of the block
                 //              |----block---|
                 //        |----range---|
-                let (left_block, right_block) = Block::split(
-                    conn,
-                    block,
-                    block.start + end - path_start,
-                    chromosome_index,
-                    phased,
-                )
-                .unwrap();
+                let split_point = block.start + end - path_start;
+                let (left_block, right_block) =
+                    Block::split(conn, block, split_point, chromosome_index, phased).unwrap();
                 Block::delete(conn, block.id);
-                // let right_block = Block::create(
-                //     conn,
-                //     &block.sequence_hash,
-                //     block_group_id,
-                //     end - path_start,
-                //     block.end,
-                //     &block.strand,
-                // );
-                // // what stuff went to this block?
                 new_edges.push((Some(new_block_id), Some(right_block.id)));
-                // let last_node = dfs.next(&graph);
-                // if last_node.is_some() {
-                //     let next_block = blocks.get(&(last_node.unwrap() as i32)).unwrap();
-                //     new_edges.push((Some(right_block.id), Some(next_block.id)));
-                // } else {
-                //     new_edges.push((Some(right_block.id), None))
-                // }
                 break;
             } else if overlap {
                 // our range is the whole block, ignore it
@@ -713,6 +704,110 @@ mod tests {
             HashSet::from_iter(vec![
                 "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
                 "AAAAAAANNNNTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_on_block_boundary_start() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 10, 10, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAANNNNTTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_on_block_boundary_end() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 9, 9, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAANNNNATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_across_entire_block_boundary() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 10, 20, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAANNNNCCCCCCCCCCGGGGGGGGGG".to_string()
+            ])
+        );
+    }
+
+    #[test]
+    fn insert_across_two_blocks() {
+        let mut conn = get_connection();
+        let (block_group_id, path_id) = setup_block_group(&mut conn);
+        let insert_sequence =
+            Sequence::create(&mut conn, "DNA".to_string(), &"NNNN".to_string(), true);
+        let insert = Block::create(
+            &conn,
+            &insert_sequence,
+            block_group_id,
+            0,
+            4,
+            &"1".to_string(),
+        );
+        BlockGroup::insert_change(&mut conn, path_id, 15, 25, &insert, 1, 0);
+
+        let all_sequences = BlockGroup::get_all_sequences(&conn, block_group_id);
+        assert_eq!(
+            all_sequences,
+            HashSet::from_iter(vec![
+                "AAAAAAAAAATTTTTTTTTTCCCCCCCCCCGGGGGGGGGG".to_string(),
+                "AAAAAAAAAATTTTTNNNNCCCCCGGGGGGGGGG".to_string()
             ])
         );
     }
